+++ apps/frontend/src/AppContent.tsx
@@ -4,29 +4,38 @@ import { GameConnection } from './views/GameConnection/GameConnection';
 import { GameReflection } from './views/GameReflection/GameReflection';
 
 const AppContent = () => {
-  const { currentView, setCurrentView } = useAppState();
+  const { currentView, setCurrentView, gameService } = useAppState();
 
   useEffect(() => {
-    gameService.gameStatus(({ status }) => {
-      if (status === 'waiting') {
-        setCurrentView(<GameConnection />);
+    const handleStatus = ({ status }: { status?: string}) => {
+      switch (status) {
+        case 'waiting':
+          setCurrentView(<GameConnection />);
+          break;
+        case  'reflection':
+          setCurrentView(<GameReflection />);
+          break;
+        default:
+          break;
       }
-      if (status === 'reflection') {
-        setCurrentView(<GameReflection />);
-      }
-    });
-  }, [setCurrentView]);
+    };
 
+    gameService.requestGameStatus(handleStatus);
+
+    return () => {
+      gameService.getSocket().off('game-status', handleStatus);
+    };
+
+    
+  }, [gameService, setCurrentView]);
+
+  
   return (
     <>
       <nav className="a11y-skip-content">
         <ul>
-          <li>
-            <a href="#main">Aller au contenu principal</a>
-          </li>
-          <li>
-            <a href="#main">Aller à un autre endroit</a>
-          </li>
+          <li><a href="#main">Aller au contenu principal</a></li>
+          <li><a href="#main">Aller à un autre endroit</a></li>
         </ul>
       </nav>
       <main id="main">{currentView}</main>
